Move {en,dis}ableFormElement to features/disable.js
This complements 2adb406df9f7d2acf4e23eea912a4900113f8b2d
