+++ src/features/disable.js
@@ -1,6 +1,6 @@
 import config from '../config'
 import { stopEverything } from '../utils/event'
-import { formElements, enableFormElement, disableFormElement } from '../utils/form'
+import { formElements } from '../utils/form'
 
 //  Replace element's html with the 'data-disable-with' after storing original html
 //  and prevent clicking on it
@@ -28,14 +28,48 @@ export function enableElement(element) {
   element.removeData('ujs:disabled')
 }
 
+/* Disables form elements:
+  - Caches element value in 'ujs:enable-with' data store
+  - Replaces element text with value of 'data-disable-with' attribute
+  - Sets disabled property to true
+*/
 export function disableFormElements(form) {
   formElements(form, config.disableSelector).each(function() {
     disableFormElement($(this))
   })
 }
 
+export function disableFormElement(element) {
+  var method, replacement
+
+  method = element.is('button') ? 'html' : 'val'
+  replacement = element.data('disable-with')
+
+  if (replacement !== undefined) {
+    element.data('ujs:enable-with', element[method]())
+    element[method](replacement)
+  }
+
+  element.prop('disabled', true)
+  element.data('ujs:disabled', true)
+}
+
+/* Re-enables disabled form elements:
+  - Replaces element text with cached value from 'ujs:enable-with' data store (created in `disableFormElements`)
+  - Sets disabled property to false
+*/
 export function enableFormElements(form) {
   formElements(form, config.enableSelector).each(function() {
     enableFormElement($(this))
   })
+}
+
+export function enableFormElement(element) {
+  var method = element.is('button') ? 'html' : 'val'
+  if (element.data('ujs:enable-with') !== undefined) {
+    element[method](element.data('ujs:enable-with'))
+    element.removeData('ujs:enable-with') // clean up cache
+  }
+  element.prop('disabled', false)
+  element.removeData('ujs:disabled')
 }
